infra_gcp/docker/requirements.txt
selenium==4.20.0
beautifulsoup4
pandas
google-cloud-bigquery
google-cloud-storage
